🔨 Refactor fortify to separate xml and fpr (#9667)
* :hammer: restructure fortify

* fortify xml

* fortify fpr

* fix imports

* flake8
